Update README.md
Corrected some wrong informations for 7-Zip. 7-Zip is also available as a 100% portable version, and it also does NOT require Admin-rights. Admin is only needed if you would like to enable the “Add 7-Zip to context menu” setting. And to be honest, theoretically it is also more platform independent than the most others because you can open ZIP as well as 7Z files on Android and on iOS too, which can't be done with Picocrypt's files.